pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "conda_sphinx_theme"
dynamic = ["version"]
description = "Conda theme for Sphinx"
readme = "README.md"
requires-python = ">=3"
license = {file = "LICENSE"}
authors = [
    {name = "Anaconda, Inc."},
]
classifiers = [
    "Framework :: Sphinx :: Theme",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Documentation",
    "Topic :: Documentation :: Sphinx",
]
dependencies = [
    "pydata-sphinx-theme>=0.13.1,<0.16.0",
    "sphinx-favicon",
]

[project.urls]
homepage = "https://github.com/conda-incubator/conda-sphinx-theme/"
repository = "https://github.com/conda-incubator/conda-sphinx-theme"

# http://www.sphinx-doc.org/en/stable/theming.html#distribute-your-theme-as-a-python-package
[project.entry-points."sphinx.html_themes"]
conda_sphinx_theme = "conda_sphinx_theme"

[tool.setuptools]
packages = ["conda_sphinx_theme"]
zip-safe = false

[tool.setuptools_scm]
version_file = "conda_sphinx_theme/_version.py"

[tool.setuptools.package-data]
conda_sphinx_theme = [
    "theme.conf",
    "*.html",
    "static/css/*.css",
    "static/js/*.js",
    "static/fonts/*.*",
    "static/*.svg",
    "static/*.ico",
    "_templates/*.html",
]
